Adds a frame for screen tests
In order to make the writing process for screen tests easier, this
commit adds some constants and functions to the screen test file.

The constants are meant to represent object data to be put in a pixel.

The functions are for getting a list of said object data in a format
that the screen can work with, and another for converting object data
for assertion.