Changes the logic of some pixel methods to make more sense
In order to make the pixel methods easier to understand, this commit
removes any 'over engineering' built into some of the pixel methods.

The methods and changes are as follows.

change_display_to()
No longer automatically assigns the display if the input was invalid.
No longer automatically assigns the display if only the key was passed in.
No longer takes an option as an argument for the object key and number.

clear_display_data()
Added to replace the use of change_display_to(None None) as
change_display_to() no longer takes options.

insert_object()
Is now much more idiomatic.

remove_displayed_object()
Now asks if you want to reassign the pixel after removal.

Along with all of this, pixels now hold their given index in the screen.
This will help with logging and debugging.