Adds pixel tests and changes pixel api + method logic
In order to work towards full pixel test coverage, this commit adds
tests for the following tests.

change_display_to()
clear_display_data()
remove_displayed_object()

In addition to these new tests, this commit also changes some of the
methods. This is due to finding flaws in the api and overall logic of
said methods when writing tests for them.