Changes pixel data removal and moves pixel methods
In order to clean up the pixel file, this commit moves a majority of the
pixel methods to separate files and implements them as public traits for
the pixel struct.

This commit also removes a redundant method that was the same as
remove_displayed_object(). With this removal a new method was added to
remove data from the pixel with a passed in object name and assigned number.