Adds tests for remove_object and changes some logic/API
In order to get full coverage of the remove_object method, this commit
adds tests for every possibility when using said method.

In addition, this commit changes some of the logic and API of some
methods. Those methods are as follows.

display()
No longer gets the first available object to display when the input is
invalid.

remove_object()
Now returns a KeyAndObjectDisplay instead of an AssignedObject

contains_assignment()
New method added to check if a specific assignment exists.

These changes will allow said methods to be more transparent in what
they do and return.