Implements the screen_printer crate
In order to render the screen cleaner and more efficiently, this commit
implements the screen_printer crate for properly rendering the screen.
This will fix the bug where too many screen updates would cause
the screen to print so fast the terminal couldn't keep up.
Along with that, printing the screen will be much more efficient than
having to print everything every frame which also clogged up the
terminal.