Implements a config file and logging to files.
In order to better debug any issues that occurr, this commit implements
the ability to log to a file. The files will be ordered by date, and logs
will be formatted with time (in UTC), file called, and line called.

In addition, this commit implements a config file. This will allow us to
more easily input preset values such as log level, clock tickrate, and
default empty pixels.