Initial implementation of objects
In order to get a baseline for how the engine will work post overhaul,
this commit removes all of the old logic pertaining to objects and
pixels.
In addition, this commit implements the Object trait. The object trait
will work with the ObjectData struct which houses all data needed for an
object to exist.