Implements a way to print objects on the screen.
In order to get an idea for how the screen will be printed, this commit
implements the basic code for getting object printed onto the screen.

There's still issues with the current implementation.
First, the air characters are just being replaced with the assigend
air character, and being printed whole. This is a problem
obviously for objects erasing anything around them with their air
characters.
Second, once an object is put in the screen it cannot be changed. This
is because objects are stored as mutable references. The fix for this
would be to wrap object_data in a Mutex wrapped in an Arc. This will
allow for a thread safe way to edit the object data.