Moves the printer to this repo and changes object storage.
In order to log the internal data of the printer, this commit merges the
project into this one. This will allow for adding any required logging
to the files for debugging.

In addition, this commit changes how object data is stored to an Arc
Mutex. This will allow for thread safe changes to be made to objects
whenever necessary.