Adds ScreenData tests and fixes a bug with objects
In order to get full coverage of the project, this commit adds tests
for a few functions in the screen_data file.

In addition, this commit fixes a bug where when moved, an object's top
left position was being calculated off of the old center position and
not the new one. This caused movements to seem delayed by 1 action every
time the object moved.