Implements object hitboxes and tells what objects collided.
In order to get objects to know what other object's hitboxes it's
currently in contact with, this commit implements everything to create
a hitbox, and check if it's inside another one.
To do this, the idea of a fully dynamic hitbox was scrapped. Instead,
hitboxes will only be rectangular in shape. This is to simplify the
checks for object collisions.

In addition, objects now hold a reference to the "Objects" struct, which
was originally only held by the screen. This is so objects know what
other objects exist, and so objects can fix what their assigned strata
is when they change it.
Along with that, this change allows us to return a copy of the ObjectData
for every object that collides with a given checked object.
This is check will return what objects it collided with, and is run
every time an object moves.
To confirm, this check will only compare the object's self with every
object that exists, no other objects will be compared with each other
through this check.