Fully implements a parser for model files
In order to have an easier and more clear way of creating model_data,
this commit implements a parser for model files.
The files will be labeled with the extension ".model".

The format of the files are as such.
Header
Data
Spacer

The header would be something such as "Skin" to say that the following
data is to be parsed as skin data.

The Skin data row has a left and right side, with the right side
indicating contained data. The sides are separated with "=". The left
side will be one of a few specific name to indicate the different fields
needed to create a ModelData struct.
The right side will contain the actual data for the field on the left.
The data will be held as such: 'data'.

The spacer is to separate different headers. This will be marked with
"--".

For things such as the model's hitbox or appearance. There's two headers
specifically for that. Instead of left and right sides to indicate data,
the entire string for the appearance and or hitboxes will be placed
under their respective headers.