Changes the way the model file parser handles Skin data
In order to improve the appearance of the parser, this commit changes
how skin data is handled. The parser use to build out a string character
by character stopping at '=', then gather the data on the right hand
side of the '='. Instead, now it splits at '=' (returning an error if
there was none), and does all required checks of the left and right hand
sides from there.