Changes any mention of the center of a model to anchor or world_anchor
In order to have a more consistent meaning as to what the center was,
this commit changes any mention of "center" to "anchor" or "world_anchor".

When talking about the character that dictated an achor from the hitbox
to the skin, "center" was replaced with "anchor".
When talking about the relative distance of the anchor to a model's top
left index in a frame, "center" was changed to something along the lines
of "world_placement_anchor".