Implements a starter for the printer to prevent a crash
In order to prevent a crash, this commit forces the user to call a
'start_printer()' method that will set the printer's origin.
If the user attempts to start the input thread before starting the
printer, an error will be returned instead.

In addition, this commit adds some documentation to the printer on how
methods can return errors.