Changes the ModelData to hold an Arc Mutex to the internal data
In order to make the api for handling model_data better, this commit
moves all data from ModelData to InternalModelData. InternalModelData
will be wrapped in an Arc Mutex internally. This will allow for 1 line
method calls on any instance of method_data without having to get any
locks on it.