Implements a teleporter model for testing the api
In order to get a feel for the new api, this commit implements a new
model named TeleportPad. This model is connected to another model of the
same time, the check for a square uses `move_to()` to 'teleport' the
square to the other TeleportPad.

Through this implementation, a bug was found with the `move_to()` method.
The method was directly using the input from the user, (based on
anchored position), when it needed to be converted to the top left
position instead.