Implements a way to remove a model from the internal storage
In order to have a way to remove internally stored models, this commit
adds the ``remove_model()`` method to ScreenData and InternalModels.
This method will remove any mention of a model of the passed in hash.
If the model never existed internally to begin with, it just returns
nothing.
If the model did exist, it returns the ModelData that was held
internally.