Adds test model and guts model_data tests
In order to get a basis on the new testing format, this commit removes
all outdated tests from the model_data test file.

In addition, this commit adds a test_square model file. This model will
be the baseline for checking if any data returned from methods is what
it's expected to be.