Removes the internal Skin in Sprite and removes outdated tests
In order to reduce confusion when dealing with the sprite, this commit
removes the skin struct entirely, and just moves all the internal data
to the sprite.

In addition, to make it easier to write tests for everything, this
commit removes any outdated tests that were found when doing the
aforementioned changes.