Adds hitbox tests and fixes a bug with hitbox anchors
This commit adds tests for most hitbox methods.

Through creating these tests, it was found that the hitbox anchors were
being calculated wrong. This didn't result in any unexpected behavior,
it just meant the hitboxes were "living in a different world" from
objects when it came to their labeled positions.