Adds all possible tests for general_data.
This commit adds tests for coordinates under the general_data file.

As I'm aware, there's no way to test the other files under general_data.
It would all require modifying the config file, which is impossible,
since tests are forced to run on the config defaults.
Any test that wouldn't require modifying the config would require a
hold on stdin and stdout, which are blocked when running tests.