Implements a method for checking model collisions for a given movement
In order to have a more consistent way to handle model collisions, this
commit implements the ``move_to_collision_check`` and
``move_by_collision_check`` methods to models.
This method will take a given movement of the corresponding movement
type. (move_to and move_by)
The method returns the collisions that would've happened had the
model made that movement.

This commit also implements those changes into the example with a new
way of dealing with model collisions.