Hides internal methods, changes how animations are handled, moves tests.
In order to not confuse any users on what they are and aren't allowed to
use, this commit restricts any internal public methods to only be usable
in the crate. This should also help with readability on the
documentation for docs.rs.

This commit also moves any internal tests inside their respective files,
and any tests dealing with the API to the tests/ directory. This should
help with understanding where/how certain methods should and shouldn't
be used in/for the project.

Finally, this commit changes how animation frames are handled by
ModelAnimators in the animation thread. Instead of storing counters for
frame iterations, it's all handled by an Iterator made for
AnimationFrames.