Removes old printer, updates animation logic, and more
In order to improve the stability of how the screen is printed, this commit
updates the printer crate to the latest version and removes any of the
old wrapper methods that improved stability.

In addition, this commit changes some of the animation logic, although
still contains some debugging lines for future tests to be added.

Finally, this commit contains now outdated code written for parsing
animation files. It was found that the readability and clarity of
creating models and animations models was just too bad. Therefore all
of that work is being pushed off to a separate program dedicated to
making models, and animations for those models.