Sets up most data strctures in model_data_structures
In order to allow separate components of the graphic engine to be able
to interact with the same data structures, this commit begins migration
of all data structures pertaining to models into its own crate.

In addition, this commit adds an `engine_math` crate. This will handle
anything pertaining to maths such as coordinates, hashing, and
rectangles.