Adds collision check with movement and tests for it
In order to allow a check for collisions with any given movement, this
commit adds the method `check_if_movement_causes_collisions`. This
method will check if a given movement will cause a collision, and
returns the list if it does. However, this method will not actually move
the model, unlike the similar `move_model` method.