Implements animation methods for model animator.
In order to have an easier access to the API of a model's animation
data, this commit implements methods that handles all of that through
the ModelManager. This commit also adds tests for all 4 methods that
were added.

In additon, this commit adds a list of `TODO`s on methods that return a
Result but haven't documented the errors that can be returned.