Adds the new implementation for ModelAppearance
In order to have a more memory efficient and logical way of obtaining a
model's appearance, this commit adds the base implementation for the
ModelAppearance struct for ModelData.

The ModelAppearance struct will contain all data for all possible
appearances of a model. The ModelAppearance will return a reference to a
Sprite contained inside. This Sprite will be calculated based on the
current state of the ModelAppearance.
The state can change based on animations running internally through the
ModelAnimator, or if the ModelAppearance has been told to set on a
resting frame.

The ModelAnimator contains a tracker for time passed. It will use this
to calcuate the current appearance of the model based on running
animations, returning a reference to the Sprite.

This commit does not fully implement being able to use this new
implementation of ModelApearance.
This commit also breaks a lot of tests for now because no implementation
of serializing and deserializing this data exists yet.