Fixes ModelAppearance and ModelAnimator bugs
In order to allow returning references to Sprites through the
get_current_appearance method, this commit changes which field under the
ModelAppearance data structure is wrapped in a RefCell.
Previously, the ModelAppearance contained a RefCell<ModelAnimationData>.
This caused problems because you can't return references from a
RefCell, in which the ModelAnimationData is used to store all animation
data. Now the ModelAnimator contained in the ModelAnimationData is
wrapped in a RefCell as that's the only part of the entire chain that
requires changing when calling the get_current_appearance method.

In addition, this commit fixes a bug for ModelAnimator's
step_animation_queue method. If this method was called when the current
animation was finished, the remaining time would be lost. That is no
longer the case.