Implements a way for the ModelManager to track collision events
In order to diversify the ways of obtaining collision events, this
commit adds a collision event tracker to the ModelManager. This will
store every collision that happened in a list ordered by when it
happened. This list contains both the ModelCollisions object, and a
timestamp for when that event occurred.