Implements an example of using the new collision event tracker
In order to test the API of a collision event tracker, this commit
implements the collision event tracker in the animation_tests example.

This example runs a method that checks all collision events that
occurred after every frame, and acts upon them. Currently, the only
thing that happens is running an animation for the AnimationActors if
anything collided into it.